macOS Functionality Partially Restored
At the end of the day I decided the .app should just be a bootstrap. The .app file also contains a unix executable that has Malice, and also contains the resources for Malice. This makes things a lot less of a headache when linking GTK.

Of course, GTK still isn't statically linked, so I will have to work on that.